requirements.txt
# Core Framework
streamlit>=1.28.0

# SOAP Client
zeep>=4.2.1

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0

# Date/Time
python-dateutil>=2.8.2

# Logging
colorlog>=6.7.0

# Type Checking (Development)
mypy>=1.5.0
types-requests>=2.31.0
